Extract files, add seq support
