+++ dotnet.ecosystem/dotnet.ecosystem.csproj
@@ -15,6 +15,7 @@
     <PackageReference Include="Neo4j.Driver" Version="5.9.0" />
     <PackageReference Include="serilog" Version="2.12.0" />
     <PackageReference Include="Serilog.Sinks.ColoredConsole" Version="3.0.1" />
+    <PackageReference Include="Serilog.Sinks.Seq" Version="5.2.2" />
     <PackageReference Include="spectre.console" Version="0.47.0" />
     <PackageReference Include="Spectre.Console.Cli" Version="0.47.0" />
     <PackageReference Include="System.CommandLine" Version="2.0.0-beta4.22272.1" />
